requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
cachetools>=5.0.0
base58>=2.1.0
psutil>=5.9.0
urllib3>=1.26.0